Move bulk update to service, refactor
